[🔧수정] 스크롤제거 코드 수정
