+++ src/App.tsx
@@ -1,14 +1,11 @@
 import { useEffect } from 'react'
-import { useRecoilValue } from 'recoil'
 import { Routes, Route } from 'react-router-dom'
 import { useKeycloak } from '@react-keycloak/web'
-import { currentModalAtom } from '@/atoms'
 import { useCreateUserCookie } from '@/hooks/_index'
 import { Intro, Home, Detail, School, Error } from '@/pages'
 import { getUserAuthCookie } from './utils'
 
 function App() {
-  const currentModal = useRecoilValue(currentModalAtom)
 
   const { keycloak } = useKeycloak()
   const createUserCookie = useCreateUserCookie()
@@ -20,15 +17,6 @@ function App() {
       createUserCookie()
     }
   }, [keycloak.authenticated])
-
-  /* 모달창이 열려있으면 스크롤바를 제거함 */
-  useEffect(() => {
-    if (currentModal) {
-      document.body.classList.add('no-scroll')
-    } else {
-      document.body.classList.remove('no-scroll')
-    }
-  }, [currentModal])
 
   return (
     <div className="app">
